% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarflowr-package.R
\docType{package}
\name{tarflowr-package}
\alias{tarflowr}
\alias{tarflowr-package}
\title{tarflowr: High-level Orchestration of 'targets' Pipelines}
\description{
Provides a simple, high-level interface for creating and executing 'targets'-based workflows, so users can focus on their specific analysis, not on pipeline orchestration. The philosophy of tarflowr is based on the concept of arbitrary "work units." This package is experimental.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brownag/rgeomorphon/}
  \item \url{https://humus.rocks/rgeomorphon/}
  \item Report bugs at \url{https://github.com/brownag/rgeomorphon/issues}
}

}
\author{
\strong{Maintainer}: Andrew Brown \email{brown.andrewg@gmail.com} (\href{https://orcid.org/0000-0002-4565-533X}{ORCID})

}
\keyword{internal}
