% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarflowr.R
\name{tarflowr_load}
\alias{tarflowr_load}
\title{Load a target from a tarflowr project into an environment.}
\usage{
tarflowr_load(name, project_dir, envir = parent.frame())
}
\arguments{
\item{name}{The unquoted name of the target to load.}

\item{project_dir}{The path to the tarflowr project directory containing
_targets/objects folder to load from.}

\item{envir}{The environment to load the target into. Defaults to the calling
environment (\code{\link[=parent.frame]{parent.frame()}})}
}
\value{
The object \code{name} is loaded into the specified environment \code{envir} as
a side-effect.
}
\description{
A simple wrapper around \code{targets::tar_load()}. This function loads the value
of a target from a completed or running tarflowr pipeline into the specified
R environment.
}
\examples{
\dontrun{

td <- file.path(tempdir(), "_my_project")

res <- tarflowr_run(
  work_units = list(a = 1, b = 2),
  process_func = function(x) x * 10,
  project_dir = td
)

# load target object into current environment as side-effect
tarflowr_load(work_seq, project_dir = td)

work_seq
}
}
\author{
Andrew G. Brown
}
