% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarflowr-orchestrator.R
\name{tarflowr_project}
\alias{tarflowr_project}
\title{Define a targets pipeline as a tarflowr work unit.}
\usage{
tarflowr_project(project_dir, result_target, name = NULL)
}
\arguments{
\item{project_dir}{The path to the directory containing the sub-project's
\verb{_targets.R} file.}

\item{result_target}{The unquoted name of the target within the sub-project
that should be returned as the result of this work unit.}

\item{name}{(Optional) A short, descriptive name for this work unit, which
will be used for naming branches in the main \code{tarflowr} pipeline. If
\code{NULL}, the basename of the \code{project_dir} will be used.}
}
\value{
A structured list object (\code{tarflowr_work_unit}) containing the
configuration for the sub-project.
}
\description{
This is a work unit constructor designed for hierarchical workflows. It
creates a configuration object that describes a sub-project to be run as a
single step within a larger \code{tarflowr} meta-pipeline.
}
\details{
The object created by this function is intended to be passed as an element in
the \code{work_units} list to \code{tarflowr_run()}, with \code{tarflowr_run_subproject()}
set as the \code{process_func}.
}
\author{
Andrew G. Brown
}
